artifacts/api-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pandas==2.2.3
rapidfuzz==3.10.1
psycopg2-binary==2.9.10
pydantic==2.10.4
python-dotenv==1.0.1
openai==1.58.1
httpx==0.28.1
